fix memory corruption caused by Int32 vs Int64 in executeImmediate calls. (cherry picked from commit dd2ad5c59c85882466c1c170e0edd6527a6d13a8)

+++ core/sql/sqlcomp/CmpSeabaseDDLcommon.cpp
@@ -9690,7 +9690,7 @@ Lng32 cliRC = 0;
                fromOwnerID);
                
 int32_t length = 0;
-int32_t rowCount = 0;
+Int64 rowCount = 0;
 
    cliRC = cliInterface.executeImmediate(buf,(char*)&rowCount,&length,NULL);
